Fixes? patient ID bug (NOW? int instead of float)

+++ mongoSetup.py
@@ -2,7 +2,7 @@ from pymodm import MongoModel, fields, base
 
 
 class Patient(MongoModel):
-    ID = fields.FloatField(primary_key=True)
+    ID = fields.IntegerField(primary_key=True)
     attending_email = fields.CharField()
     age = fields.FloatField()
     heart_rates = fields.ListField()
